requirements.txt
streamlit==1.31.0
crewai==0.11.0
langchain==0.1.7
openai==1.12.0
unstructured==0.10.30
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0
